# Conflicts:
#	src/main/java/com/azid/auth/backend/AZ/Auth/controller/ClaimController.java
#	src/main/java/com/azid/auth/backend/AZ/Auth/dto/ClaimResponseDto.java
#	src/main/java/com/azid/auth/backend/AZ/Auth/model/Payment.java
#	src/main/java/com/azid/auth/backend/AZ/Auth/model/Policy.java
#	src/main/java/com/azid/auth/backend/AZ/Auth/repository/ClaimTypeRepository.java
#	src/main/java/com/azid/auth/backend/AZ/Auth/repository/UserRepository.java
#	src/main/java/com/azid/auth/backend/AZ/Auth/service/ClaimService.java
#	src/main/java/com/azid/auth/backend/AZ/Auth/service/UserService.java

+++ src/main/java/com/azid/auth/backend/AZ/Auth/model/Plan.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -24,12 +26,12 @@ public class Plan {
     private Double coverageAmount;
 
     @Column(name = "base_premium")
-    private Double basePremium;
+    private BigDecimal basePremium;
 
     @Column(name = "duration")
     private Integer duration;
 
     @Column(name = "status")
-    private String status;
+    private String status; // ACTIVE / INACTIVE
 
 }
